Update ProductCipher.Java
here during encryption and decryption the n value should be the same therefore decryption of the pain text was correct. Didnt quite figure out the encryption with the user input value n , i will do it in sometime when i understand the bug. Would be great if you merge this by then.